Show image results from each step of the algorithm
The Jupyter notebook now displays a subplot of all test images, before
and after each step of the algorithm. A demo_show() function was
implemented to achieve this (keeping track of all the active figures)
without messing up the code too much. This functionality is controlled
by a boolean called demo_mode, and by setting this to false before the
video processing loop, all figures will be prevented and the the code
will be almost as fast as without this demo functionality.