Oversample images to get equal numbers of all labels
Just duplicate images of less represented classes in the training
examples to get an equal number of all class labels.

Also implement class weights which is another method I tried to punish
well represented class labels in favor of not so frequent ones. I was
however not able to achieve any good result with this method.