Use left and right cam only when driving straight
For now, only use left and right camera to stabilize when driving
fairly straight in the normal and reversed driving sequences.

This because it was hard to find a left and right camera steering
offset that actually makes the result better compared to when only
using the center camera. It might be that the steering offset
shouldn't be the same in a sharp curve, when doing a recovery
action and/or when driving straight.