Fix usage of Keras fit_generator
The fit_generator works differently in Keras 2 and doesn't seem to be
backwards compatible with Keras 1, this although it only outputs a
warning suggesting that the arguments should be updated.

What I think is happening is that one batch of samples is generated
every time the fit function asks for a single sample, so I guess the
result is as if the whole data set was oversampled by a factor of
batch_size and the actual batch size applied was batch_size^2 (but
I'm not sure about this).

This explains why the trainings so far have taken so much time and
that the result overfits already at second or third epoch.