Tuned steering offset for left and right camera
Update default steering offset to 0.0245 after manual inspection of
the outcome from training with different offset values.

What I was looking at when trying out different offset values was
that the vehicle shouldn't wobble when driving straight, which it
clearly does for offsets larger than ~0.026. On the other hand, the
vehicle shouldn't be to satisfied when drifting away from the center
of the road which seems to be the case when the offset drops below
~0.02. Too small offset also seems to make the model instable, such
that it change the output a lot from frame to frame.

The bash script trainer.sh contains the model configurations tried.