Train on data from both track 1 and 2
Data from track 2 is labeled by which lane it's sampled. The model is
trained to make the vehicle drive in the right lane on track 2.
Samples from left lane are still used but augmented by horizontal
flip to make it look like driving in the right lane from a different
but similar track.