Augment image shadows
A shadow is randomly generated and drawn into the image. This by
dividing the image into two parts by a straight line, where one side
is darkened. The darkening is done without affecting the colors very
much, which is achieved by first converting color-space from BGR to
HSV, and then only modify the V-channel. The image is then converted
back to BGR.