Use same colorspace for driving and training
Noticed that the drive.py file inputs RGB images to the model, which
however is trained with BGR. It's somewhat a surprise that the model
did so well even with a mix up of blue and red channels. Perhaps most
of the information used by the model is from the green channel.

But although the model already did work to some level, this bug-fix
sure gives a huge improvement.