Helper tool to find perspective transform
Implements an interactive tool to let the user select perspective
transformation source points in an image, and then measure dash line
length and lane widths in the transformed image. The measurements
are used to adjust the transform such that 30 meters is visualized in
front of the vehicle and 5 meters left and right. The measurements
also provides estimates of pixel resolution in meters, which will be
used later on when estimating lane curvature.