Classifier processing search windows
Implements a classifier module which use the trainer module to train
a classifier and then use it to predict the class of image patches.

The trained classifier is stored to file, from which it's loaded in
the next execution, e.g. the training is only done once.

The classifier outputs a confidence score, i.e. the distance of the
sample to the hyperplane. A colormap is defined to indicate this
value when plotting boxes around detections.