Tunes the grid generator parameters
Improves the camera parameters used to better fit the Video. The
previous parameters did look somewhat OK in a lesson test image,
but apparently that was another camera/mount.

Found information about which camera that (probably?) is used from
the Udacity camera-mount github repo, and could then look up actual
FOV parameters in lens specification.

https://github.com/udacity/camera-mount
http://www.fujifilmusa.com/products/optical_devices/machine-vision/1-15-mp/cf125ha-1/

By setting the search corridor width to the same as the lane, 3.6576
meters (12 feets) and the distances between grids to align with the
dash and gaps lengths, (10 resp. 30 feets), it was possible to
manually calibrate the camera parameters z, pitch and yaw, such that
the result looks reasonable OK for the project_video.mp4.

I also decided to go with a conic search corridor such that the
search area is wider further away. This mainly because the
positioning of the grids are more uncertain further away (no path
prediction available), and it's also of more interest to find objects
not directly at the path at some distance since those objects might
change position until the ego vehicle reach there.