Tool to manually label extracted detections
Iterates directory of file samples from extracted detections, and let
user sort them into TP, FP, or unspecified using the left and right
arrow keys. TP and FP samples are then copied to separate
directories. It's possible to step forward and backwards and redo
previous choices also using the number keys. Current label of an file
is shown as a green rectangle for TP and red cross for FP.

The purpose of this tool is hard negative mining, e.g. only use it to
label FP patches.