Training data extraction tool updated to handle dataset 2
Updates the tool to also handle the second set of labeled training
data provided by Udacity. The main difference is this set is that
there are more label types and also an occluded flag. I choose
to put occluded objects into a different folder to be able to skip
them in training, or control when during the training they are used.

https://github.com/udacity/self-driving-car/tree/master/annotations