Jacobian matrix calculation - handle division by zero
If both px and py is zero, then several entries of the Jacobian matrix
will end up having NaN assuming the compiler is implemented according
to IEEE 754. In other words, there will not be a crash in the Jacobian
matrix calculation due to division by zero, but the NaN values in the
Jacobian matrix will persistently end up in x_state estimation and the
uncertainty covariance matrix P. To avoid this, I've chosen to just
set all values to zero, which effectively ignores this measurement
(relying on lidar and/or state prediction to move the estimate away
from origin).

Note that this scenario is highly unlikely when tracking an object
since it actually implies that the tracked object has collided with
ego vehicle.

Some minor code performance optimizations are also implemented by reusing
calculations made for multiple Jacobian matrix entries.