Catch the run away car
Predicts the position of the run away car ahead of time based on
current Kalman filter state and the assumption that the car has
constant velocity and yaw rate. The prediction is repeated for every
delta_t, until finding a prediction that the hunter will be able to
intercept. This is then repeated whenever the Kalman filter state is
updated.