Particle filter data association
Associate observations to map landmarks by using the nearest-
neighbor approach.

The proposed interface was changed such that observations are
transformed to map coordinates instead of the other way around.
I think doing the transformation in this direction feels more
intuitive and it will probably also result in less computations
assuming that there are more landmarks than observations in a real
application.