Steering and throttle control
First implementation of PID controller for steering, and also for the
throttle. The steering control parameters are just selected as in an
unrelated lesson example, but at least it seems to make the vehicle
drive around the track at 30 MPH without crashing. The throttle is
easier to control (at least on this flat track), and the selected
parameters seems to do the job just fine.