Compensate for latency
This change activates the delay that simulates latency in the system,
which is compensated for by predicting the vehicle state at the point
in time when the next actuations can be expected to kick-in.

To make this work with the basic motion model, some tuning was also
necessary that essentially force the controller to not turn too much.