Extend previous coordinate vector
Extend previous coordinate vector instead of recalculating in order
to not confuse the simulator when applying the next vector. At least
I think that's the problem causing sudden acceleration every now and
then in previous commit.

This commit is very much inspired by the QA session for this
project, also utilizing the same approximation for the step length
when calculating the extended coordinates.