Implement optimal acceleration
Calculates the path to reach target speed as fast as possible without
exceeding acceleration and jerk limits.

Seems like the simulator doesn't complain much about exceeding the
jerk limit, which probably is due to that the number of significant
digits in the websocket interface is limited to 7. This commit still
calculates the acceleration to be within the jerk limit, but it's
actually possible to increase the jerk limit to at least 100 m/s^3
without any warnings about exceeding the jerk limit from the
simulator.