More advanced path prediction for other vehicles
Try to predict lane changes by looking at the vehicles position
within the lane and its velocity in d direction. Since the Frenet
velocities is unknown, these values are calculated based on vx, vy,
which is converted to Frenet by transforming multiple coordinates
along the Cartesian velocity vector to Frenet coordinates and then
averaging the delta distances. The averaging is done as attempt to
overcome problems with the Frenet transform , but the result is still
unreliable. A sanity check is implemented to indicate if there was
problems with the transform, with a fallback to predict the
trajectory using vx/vy.