Trajectory intention and target lane
Store intended path and target lane number for each coordinate in the
trajectory, such that this can be used in cost functions to reduce
the number of intention changes.

Since the trajectory feedback from the simulator not contains this
information, it was necessary to create a local copy of previous
trajectory that is adjusted at start of frame to remove all
coordinates that already has been processed, such that it matches the
previous trajectory from the simulator.