Output full trajectory to simulator
The full trajectory is displayed in the simulator, but only a few
coordinates are kept from previous output when updating the
trajectory. This is to better understand the intention of the
output trajectory, in particular when the algorithm becomes uncertain
if it should do a lane change or not which may produce a wobbly path.