Fix bug in intentions cost function
The trajectory intention was set to None in the later part of the ego
vehicle trajectory. This made the cost of changing intention after
zero, even if that's changing lane. The intention is now instead set
to Keep lane.

Also add intention when applicable in prediction of other vehicles,
just for debug purpose.