Cost function for faster vehicles from behind
Add a fixed cost for trajectories changing lane where there are other
vehicles intercepting ego within 3 seconds.

With this change, it's possible to make the cost function for
changing lane in front of others less restrictive, permitting changes
even if the distance not is more than 0.25 seconds. This allows for
smoother take overs of slower vehicles.