Reduce max lane change distance
There was a lane change that simply took too long time, e.g. without
the algorithm changing it's intention back and forth. To counter
this behavior, the lengths of generated trajectories are reduced, in
particular for Keep Lane intention, which takes over in the final
part of a lane change (assuming the trajectory is recalculated before
a lane change is done).