Make sanity check for frenet coordinates less strict
The predictions of other vehicles rely on vx,vy when the Frenet
coordinates are found to be bad in some way, typically due to large
difference in velocity when comparing vx, vy with calculated vs, vd.

The predictions made from vx, vy are however really bad when going
through curves. It might for example suggest that the vehicle driving
in the next lane will cross ego lane and then continue far of the
road, which might reject the most reasonable trajectories of the ego
vehicle due to the collision cost function.

This commit makes the sanity check less strict, since the prediction
using somewhat incorrect Frenet coordinates will probably be better
than the fall-back prediction.