Decrease critical distance to vehicle ahead
Decrease what's considered to be a critical distance to vehicle ahead
from 1.5 to 1.0 seconds.

Also remove the recommended distance cost function, since 3 seconds
barely makes sense for Swedish roads where vehicles keep to the right
and let faster vehicles pass to the left (at least on paper). In
American style traffic as in the simulator it gets impossible to find
a path forward when following this recommendation.