Include more vehicles behind in calculation distance ahead
Vehicles behind ego are skipped due to bad predictions, but it's
necessary to include the ones just behind ego vehicle to not get lane
change trajectories that intercept a vehicle passing by. This is by
no means a perfect solution, e.g. it would be better to also consider
the speed of passing vehicles, and even better to fix the predictions
to completely avoid the need for this type of work-arounds.
.